Int to Float change

+++ src/main/java/reports/EmployeeAlphabeticalReport.java
@@ -13,12 +13,12 @@ public class EmployeeAlphabeticalReport  {
     private static ArrayList<ProjectTask> projectTasks = Menu.getProjectTasks();
     public static void printReport(int year){
 
-        Map<String, Integer> map = new HashMap<>();
+        Map<String, Float> map = new HashMap<>();
         for (ProjectTask p: projectTasks) {
 
             if (p.getDate().getYear() == year) {
 
-                int hours;
+                float hours;
                 if (map.get(p.getEmployeeName()) == null) {
                     hours = 0;
                 } else {
@@ -27,7 +27,7 @@ public class EmployeeAlphabeticalReport  {
                 map.put(p.getEmployeeName(), hours + p.getHours());
             }
         }
-        Map<String, Integer> result = new TreeMap<String, Integer>(map);
+        Map<String, Float> result = new TreeMap<String, Float>(map);
 
         System.out.println("Wyświetlenie raportu godzin pracownikow w danym roku:");
